+++ src/main/java/com/dev/taskflow/Service/TaskService.java
@@ -5,6 +5,7 @@ import com.dev.taskflow.DTOs.TaskDTO;
 import com.dev.taskflow.DTOs.TaskUpdateDTO;
 import com.dev.taskflow.Entity.Category;
 import com.dev.taskflow.Entity.Task;
+import com.dev.taskflow.Entity.User;
 import com.dev.taskflow.Enums.TaskStatus;
 import com.dev.taskflow.Repository.CategoryRepository;
 import com.dev.taskflow.Repository.TaskRepository;
@@ -18,6 +19,9 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+
+import static com.dev.taskflow.Util.SecurityUtil.getAuthenticatedUser;
+
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
@@ -28,7 +32,9 @@ public class TaskService implements ITaskService {
 
     @Override
     public Page<TaskDTO> getTasks(String title, TaskStatus status, Long categoryId, Pageable pageable) {
-        Specification<Task> spec = ((root, query, cb) -> cb.conjunction());
+        User user = getAuthenticatedUser();
+
+        Specification<Task> spec = TaskSpecification.hasUser(user);
 
         if (title != null && !title.isBlank()) {
             spec = spec.and(TaskSpecification.titleContains(title));
@@ -39,9 +45,13 @@ public class TaskService implements ITaskService {
         }
 
         if (categoryId != null) {
-            Category category = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new EntityNotFoundException("Categoria não encontrada com id: " + categoryId));
-            spec = spec.and(TaskSpecification.hasCategory(category));
+            var category = categoryRepository.findById(categoryId);
+
+            if (category.isEmpty()) return Page.empty();
+
+            if (!category.get().getUser().getId().equals(user.getId())) return Page.empty();
+
+            spec = spec.and(TaskSpecification.hasCategory(category.get()));
         }
 
         return taskRepository.findAll(spec, pageable)
@@ -50,8 +60,14 @@ public class TaskService implements ITaskService {
 
     @Override
     public TaskDTO getById(Long id) {
+        User user = getAuthenticatedUser();
+
         Task task = taskRepository.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada com id: " + id));
+            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada"));
+
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("Tarefa não encontrada");
+        }
 
         return toDTO(task);
     }
@@ -59,11 +75,18 @@ public class TaskService implements ITaskService {
     @Override
     @Transactional
     public TaskDTO createTask(TaskCreateDTO dto) {
+        User user = getAuthenticatedUser();
+
         Task newTask = new Task(dto.title(), dto.description());
+        newTask.setUser(user);
 
         if (dto.categoryId() != null) {
             Category category = categoryRepository.findById(dto.categoryId())
-                .orElseThrow(() -> new EntityNotFoundException("Categoria não encontrada com id: " + dto.categoryId()));
+                .orElseThrow(() -> new EntityNotFoundException("Categoria não encontrada"));
+
+            if (!category.getUser().getId().equals(user.getId())) {
+                throw new SecurityException("Categoria não encontrada");
+            }
 
             newTask.setCategory(category);
         }
@@ -75,8 +98,14 @@ public class TaskService implements ITaskService {
     @Override
     @Transactional
     public TaskDTO updateTask(Long id, TaskUpdateDTO dto) {
+        User user = getAuthenticatedUser();
+
         Task task = taskRepository.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada com id: " + id));
+            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada"));
+
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("Tarefa não encontrada");
+        }
 
         task.setTitle(dto.title());
         task.setDescription(dto.description());
@@ -87,16 +116,29 @@ public class TaskService implements ITaskService {
     @Override
     @Transactional
     public void deleteTask(Long id) {
+        User user = getAuthenticatedUser();
+
         Task task = taskRepository.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada com id: " + id));
+            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada"));
+
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("Tarefa não encontrada");
+        }
+
         taskRepository.delete(task);
     }
 
     @Override
     @Transactional
     public TaskDTO updateTaskStatus(Long id, TaskStatus status) {
+        User user = getAuthenticatedUser();
+
         Task task = taskRepository.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada com id: " + id));
+            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada"));
+
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("Tarefa não encontrada");
+        }
 
         task.setStatus(status);
 
@@ -106,11 +148,21 @@ public class TaskService implements ITaskService {
     @Override
     @Transactional
     public TaskDTO updateTaskCategory(Long id, Long categoryId) {
+        User user = getAuthenticatedUser();
+
         Task task = taskRepository.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada com id: " + id));
+            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada"));
+
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("Tarefa não encontrada");
+        }
 
         Category category = categoryRepository.findById(categoryId)
-            .orElseThrow(() -> new EntityNotFoundException("Categoria não encontrada com id: " + categoryId));
+            .orElseThrow(() -> new EntityNotFoundException("Categoria não encontrada"));
+
+        if (!category.getUser().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("Categoria não encontrada");
+        }
 
         task.setCategory(category);
         category.addTask(task);
@@ -120,8 +172,14 @@ public class TaskService implements ITaskService {
     @Override
     @Transactional
     public TaskDTO removeCategory(Long id) {
+        User user = getAuthenticatedUser();
+
         Task task = taskRepository.findById(id)
-            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada com id: " + id));
+            .orElseThrow(() -> new EntityNotFoundException("Tarefa não encontrada"));
+
+        if (!task.getUser().getId().equals(user.getId())) {
+            throw new EntityNotFoundException("Tarefa não encontrada");
+        }
 
         Category category = task.getCategory();
         task.setCategory(null);
